Run `generate-fake-app` before running `jest`
Since the current code has a tiny race condition when creating fake apps
because `jest` runs multiple test files in parallel.

This commit introduces a new npm task to generate fake apps and
run it before running `jest` so that we can ensure that this process
runs only once.
