+++ integration-test/src/test/helpers/testAutifyCliSnapshot.ts
@@ -1,25 +1,8 @@
 /* eslint-disable unicorn/filename-case */
-import { existsSync, mkdirSync, writeFileSync } from "node:fs";
-import { join } from "node:path";
 import { execAutifyCli } from "../helpers/execAutifyCli";
 
-const iosBuildPath = "ios.app";
-const androidBuildPath = "android.apk";
-// https://commons.wikimedia.org/wiki/File:Transparent.gif
-const tinyBinary = Buffer.from(
-  "R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7",
-  "base64"
-);
-// Generate fake build files if not exiting.
-// For recording, you should setup the real files on the same paths.
-if (!existsSync(iosBuildPath)) {
-  mkdirSync(iosBuildPath); // *.app is a directory
-  writeFileSync(join(iosBuildPath, "ios"), tinyBinary); // Add a fake binary file
-}
-
-if (!existsSync(androidBuildPath)) {
-  writeFileSync(androidBuildPath, tinyBinary); // Create a fake binary file
-}
+export const iosBuildPath = "ios.app";
+export const androidBuildPath = "android.apk";
 
 export const webTestRun =
   "web test run https://app.autify.com/projects/743/scenarios/91437";
